+++ packages/crypto/src/encryption.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, test } from 'vitest';
-import { Key, KeyPrint, uintArraysEqual } from '../src/encryption';
+import { Key, KeyPrint, uintArraysEqual } from './encryption';
 import { Box } from '@penumbra-zone/types/box';
 
 // NOTE: To have the most accurate representation, these the web crypto API tests run in a browser environment
