+++ packages/crypto/src/mnemonic.test.ts
@@ -5,7 +5,7 @@ import {
   isInWordList,
   SeedPhraseLength,
   validateSeedPhrase,
-} from '../src/mnemonic';
+} from './mnemonic';
 import { wordlists } from 'bip39';
 
 describe('Mnemonic tests', () => {
